+++ plinth/modules/first_boot/templatetags/firstboot_extras.py
@@ -21,13 +21,12 @@ Template tags for first boot module.
 
 from django import template
 
-from plinth import kvstore
+from plinth.modules import first_boot
 
 register = template.Library()
 
 
 @register.simple_tag
-def firstboot_is_finished():
+def firstboot_is_completed():
     """Return whether firstboot process is completed."""
-    state = kvstore.get_default('setup_state', 0)
-    return state == 1
+    return first_boot.is_completed()