+++ plinth/modules/minetest/__init__.py
@@ -16,8 +16,10 @@
 #
 
 """
-Plinth module for minetest.
+Plinth module for Minetest server.
 """
+
+import augeas
 
 from django.urls import reverse_lazy
 from django.utils.translation import ugettext_lazy as _
@@ -28,7 +30,6 @@ from plinth import cfg
 from plinth import frontpage
 from plinth import service as service_module
 from plinth.utils import format_lazy
-from plinth.views import ServiceView
 
 
 version = 2
@@ -43,8 +44,7 @@ managed_packages = ['minetest-server', 'minetest-mod-advspawning',
                     'minetest-mod-animalmaterials', 'minetest-mod-animals',
                     'minetest-mod-mesecons', 'minetest-mod-mobf-core',
                     'minetest-mod-mobf-trap', 'minetest-mod-moreblocks',
-                    'minetest-mod-nether', 'minetest-mod-torches',
-                    ]
+                    'minetest-mod-nether', 'minetest-mod-torches']
 
 title = _('Block Sandbox \n (Minetest)')
 
@@ -56,6 +56,9 @@ description = [
           'a <a href="http://www.minetest.net/downloads/">Minetest client</a> '
           'is needed.'), box_name=_(cfg.box_name)),
 ]
+
+CONFIG_FILE = '/etc/minetest/minetest.conf'
+AUG_PATH = '/files' + CONFIG_FILE + '/.anon'
 
 
 def init():
@@ -107,13 +110,6 @@ def disable():
     frontpage.remove_shortcut('minetest')
 
 
-class MinetestServiceView(ServiceView):
-    service_id = managed_services[0]
-    diagnostics_module_name = "minetest"
-    description = description
-    show_status_block = True
-
-
 def diagnose():
     """Run diagnostics and return the results."""
     results = []
@@ -121,3 +117,50 @@ def diagnose():
     results.append(action_utils.diagnose_port_listening(30000, 'udp4'))
 
     return results
+
+
+def load_augeas():
+    """Initialize Augeas."""
+    aug = augeas.Augeas(flags=augeas.Augeas.NO_LOAD +
+                        augeas.Augeas.NO_MODL_AUTOLOAD)
+    aug.set('/augeas/load/Php/lens', 'Php.lns')
+    aug.set('/augeas/load/Php/incl[last() + 1]', CONFIG_FILE)
+    aug.load()
+    return aug
+
+
+def get_max_players(aug):
+    """Return the maximum players allowed on the server at one time."""
+    value = aug.get(AUG_PATH + '/max_users')
+    if value:
+        return int(value)
+
+
+def is_creative_mode_enabled(aug):
+    """Return whether creative mode is enabled."""
+    value = aug.get(AUG_PATH + '/creative_mode')
+    return value == 'true'
+
+
+def is_pvp_enabled(aug):
+    """Return whether PVP is enabled."""
+    value = aug.get(AUG_PATH + '/enable_pvp')
+    return value == 'true'
+
+
+def is_damage_enabled(aug):
+    """Return whether damage is enabled."""
+    value = aug.get(AUG_PATH + '/enable_damage')
+    return value == 'true'
+
+
+def get_configuration():
+    """Return the current configuration."""
+    aug = load_augeas()
+    conf = {
+        'max_players': get_max_players(aug),
+        'creative_mode': is_creative_mode_enabled(aug),
+        'enable_pvp': is_pvp_enabled(aug),
+        'enable_damage': is_damage_enabled(aug),
+    }
+    return conf