+++ plinth/tests/test_middleware.py
@@ -21,6 +21,7 @@ Test module for Plinth's custom middleware.
 
 from unittest.mock import Mock, patch
 
+from django.contrib.auth.models import AnonymousUser, User
 from django.http import HttpResponse
 from django.test import TestCase
 from django.test.client import RequestFactory
@@ -50,7 +51,8 @@ class TestSetupMiddleware(TestCase):
             'view_kwargs': {},
         }
 
-    def test_404_urls(self):
+    @patch('django.urls.reverse', return_value='users:login')
+    def test_404_urls(self, reverse):
         """Test how middleware deals with 404 URLs."""
         request = RequestFactory().get('/plinth/non-existing-url')
 
@@ -58,7 +60,8 @@ class TestSetupMiddleware(TestCase):
 
         self.assertEqual(response, None)
 
-    def test_url_not_an_application(self):
+    @patch('django.urls.reverse', return_value='users:login')
+    def test_url_not_an_application(self, reverse):
         """Test that none is returned for URLs that are not applications."""
         request = RequestFactory().get('/plinth/')
 
@@ -68,7 +71,8 @@ class TestSetupMiddleware(TestCase):
 
     @patch('plinth.module_loader.loaded_modules')
     @patch('django.urls.resolve')
-    def test_module_is_up_to_date(self, resolve, loaded_modules):
+    @patch('django.urls.reverse', return_value='users:login')
+    def test_module_is_up_to_date(self, reverse, resolve, loaded_modules):
         """Test that none is returned when module is up-to-date."""
         resolve.return_value.namespaces = ['mockapp']
         module = Mock()
@@ -85,25 +89,33 @@ class TestSetupMiddleware(TestCase):
     @patch('plinth.views.SetupView')
     @patch('plinth.module_loader.loaded_modules')
     @patch('django.urls.resolve')
-    def test_module_view(self, resolve, loaded_modules, setup_view):
-        """Test that setup view is returned."""
+    @patch('django.urls.reverse', return_value='users:login')
+    def test_module_view(self, reverse, resolve, loaded_modules, setup_view):
+        """Test that only registered users can access the setup view."""
         resolve.return_value.namespaces = ['mockapp']
         module = Mock()
         module.setup_helper.is_finished = None
         loaded_modules.__getitem__.return_value = module
         view = Mock()
         setup_view.as_view.return_value = view
-
         request = RequestFactory().get('/plinth/mockapp')
-        self.middleware.process_view(request, **self.kwargs)
 
+        # Verify that anonymous users cannot access the setup page
+        request.user = AnonymousUser()
+        self.middleware.process_view(request, **self.kwargs)
         setup_view.as_view.assert_called_once_with()
+        view.assert_not_called()
+
+        # Verify that logged-in users can access the setup page
+        request.user = User(username='johndoe')
+        self.middleware.process_view(request, **self.kwargs)
         view.assert_called_once_with(request, setup_helper=module.setup_helper)
 
     @patch('django.contrib.messages.success')
     @patch('plinth.module_loader.loaded_modules')
     @patch('django.urls.resolve')
-    def test_install_result_collection(self, resolve, loaded_modules,
+    @patch('django.urls.reverse', return_value='users:login')
+    def test_install_result_collection(self, reverse, resolve, loaded_modules,
                                        messages_success):
         """Test that module installation result is collected properly."""
         resolve.return_value.namespaces = ['mockapp']